+++ pydidas/contexts/diffraction_exp_context/diffraction_exp_context_io_base.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -19,16 +21,17 @@ DiffractionExperimentContext should inherit from.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["DiffractionExperimentContextIoBase"]
 
-from ...core.io_registry import GenericIoBase
+
 from ...core import UserConfigError
-from .diffraction_exp_context_io_meta import DiffractionExperimentContextIoMeta
+from ...core.io_registry import GenericIoBase
 from .diffraction_exp_context import DiffractionExperimentContext
+from .diffraction_exp_context_io_meta import DiffractionExperimentContextIoMeta
 
 
 EXP = DiffractionExperimentContext()
@@ -65,10 +68,17 @@ class DiffractionExperimentContextIoBase(
             raise UserConfigError(_text)
 
     @classmethod
-    def _write_to_exp_settings(cls):
+    def _write_to_exp_settings(cls, diffraction_exp=None):
         """
-        Write the loaded (temporary) Parameters to the DiffractionExperimentContext.
+        Write the loaded (temporary) Parameters to a Diffraction.
+
+        Parameters
+        ----------
+        diffraction_exp : Union[DiffractionExperiment, None], optional
+            The instance to be updated. If None, the generic
+            DiffractionExperimentContext will be used. The default is None.
         """
+        _exp = EXP if diffraction_exp is None else diffraction_exp
         for key in cls.imported_params:
-            EXP.set_param_value(key, cls.imported_params[key])
+            _exp.set_param_value(key, cls.imported_params[key])
         cls.imported_params = {}