+++ pydidas/contexts/diffraction_exp_context/diffraction_exp_context_io_yaml.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -19,21 +21,21 @@ export DiffractionExperimentContext metadata from a YAML file.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["DiffractionExperimentContextIoYaml"]
 
-from numbers import Real, Integral
+from numbers import Integral, Real
 
-import yaml
 import numpy as np
+import yaml
 
-from ...core.constants import YAML_EXTENSIONS, LAMBDA_IN_A_TO_E
 from ...core import UserConfigError
-from .diffraction_exp_context_io_base import DiffractionExperimentContextIoBase
+from ...core.constants import LAMBDA_IN_A_TO_E, YAML_EXTENSIONS
 from .diffraction_exp_context import DiffractionExperimentContext
+from .diffraction_exp_context_io_base import DiffractionExperimentContextIoBase
 
 
 EXP = DiffractionExperimentContext()
@@ -57,7 +59,7 @@ class DiffractionExperimentContextIoYaml(DiffractionExperimentContextIoBase):
         filename : str
             The filename of the file to be written.
         """
-        _EXP = kwargs.get("context", None) if "context" in kwargs else EXP
+        _EXP = kwargs.get("diffraction_exp", EXP)
         cls.check_for_existing_file(filename, **kwargs)
         tmp_params = _EXP.get_param_values_as_dict()
         # need to convert all float values to generic python "float" to
@@ -71,7 +73,7 @@ class DiffractionExperimentContextIoYaml(DiffractionExperimentContextIoBase):
             yaml.safe_dump(tmp_params, stream)
 
     @classmethod
-    def import_from_file(cls, filename):
+    def import_from_file(cls, filename, diffraction_exp=None):
         """
         Restore the DiffractionExperimentContext from a YAML file.
 
@@ -79,6 +81,8 @@ class DiffractionExperimentContextIoYaml(DiffractionExperimentContextIoBase):
         ----------
         filename : str
             The filename of the file to be written.
+        diffraction_exp : Union[DiffractionExperiment, None], optional
+            The DiffractionExperiment instance to be updated.
         """
         with open(filename, "r") as stream:
             try:
@@ -95,4 +99,4 @@ class DiffractionExperimentContextIoYaml(DiffractionExperimentContextIoBase):
             "xray_wavelength", np.nan
         )
         cls._verify_all_entries_present()
-        cls._write_to_exp_settings()
+        cls._write_to_exp_settings(diffraction_exp=diffraction_exp)