+++ tests/contexts/diffraction_exp_context/test_diffraction_exp_context.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -16,15 +18,15 @@
 """Unit tests for pydidas modules."""
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 
 
-import unittest
-import copy
 import logging
+import unittest
+from numbers import Real
 
 import pyFAI
 
@@ -61,7 +63,7 @@ class TestDiffractionExperimentContext(unittest.TestCase):
         obj.set_param_value("xray_energy", _new_E)
         self.assertEqual(obj.get_param_value("xray_energy"), _new_E)
         self.assertAlmostEqual(
-            obj.get_param_value("xray_wavelength"), 0.7897080652951567, delta=0.00005
+            obj.get_param_value("xray_wavelength"), 0.78970806, delta=0.00005
         )
 
     def test_set_param_wavelength(self):
@@ -70,7 +72,7 @@ class TestDiffractionExperimentContext(unittest.TestCase):
         obj.set_param_value("xray_wavelength", _new_lambda)
         self.assertEqual(obj.get_param_value("xray_wavelength"), _new_lambda)
         self.assertAlmostEqual(
-            obj.get_param_value("xray_energy"), 12.55345175429956, delta=0.0005
+            obj.get_param_value("xray_energy"), 12.5534517, delta=0.0005
         )
 
     def test_get_detector__from_param_name(self):
@@ -98,15 +100,21 @@ class TestDiffractionExperimentContext(unittest.TestCase):
         self.assertEqual(_det.pixel1, 1e-6 * _pixelsize)
         self.assertEqual(_det.pixel2, 1e-6 * _pixelsize)
 
-    def test_copy(self):
-        obj = DiffractionExperimentContext()
-        obj2 = copy.copy(obj)
-        self.assertEqual(obj, obj2)
-
     def test_set_detector_params_from_name__wrong_name(self):
         obj = DiffractionExperimentContext()
         with self.assertRaises(NameError):
             obj.set_detector_params_from_name("no such detector")
+
+    def test_update_from_diffraction_exp(self):
+        obj = DiffractionExperimentContext()
+        obj.set_param_value("detector_name", "Eiger 9M")
+        _exp = DiffractionExperiment()
+        _exp.update_from_diffraction_exp(obj)
+        for _key, _val in obj.get_param_values_as_dict().items():
+            if isinstance(_val, Real):
+                self.assertAlmostEqual(_val, _exp.get_param_value(_key), delta=0.000001)
+            else:
+                self.assertEqual(_val, _exp.get_param_value(_key))
 
     def test_set_detector_params_from_name(self):
         _det = {"name": "Pilatus 300k", "pixsize": 172, "npixx": 487, "npixy": 619}