+++ tests/contexts/diffraction_exp_context/test_diffraction_exp_context_io_meta.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -16,18 +18,16 @@
 """Unit tests for pydidas modules."""
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 
 
-import os
 import unittest
-import shutil
-import tempfile
 
 from pydidas.contexts.diffraction_exp_context import (
+    DiffractionExperiment,
     DiffractionExperimentContext,
     DiffractionExperimentContextIoBase,
     DiffractionExperimentContextIoMeta,
@@ -49,6 +49,7 @@ class TestIo(DiffractionExperimentContextIoBase):
         cls.exported = False
         cls.export_filename = None
         cls.import_filename = None
+        cls.diffraction_exp = None
 
     @classmethod
     def export_to_file(cls, filename, **kwargs):
@@ -56,31 +57,39 @@ class TestIo(DiffractionExperimentContextIoBase):
         cls.export_filename = filename
 
     @classmethod
-    def import_from_file(cls, filename):
+    def import_from_file(cls, filename, diffraction_exp=None):
         cls.imported = True
+        cls.diffraction_exp = EXP if diffraction_exp is None else diffraction_exp
         cls.import_filename = filename
 
 
 class TestExperimentSettingsIoMeta(unittest.TestCase):
     def setUp(self):
-        self._tmppath = tempfile.mkdtemp()
         TestIo.reset()
 
     def tearDown(self):
-        shutil.rmtree(self._tmppath)
         EXP.restore_all_defaults(True)
 
     def test_export_to_file(self):
-        _fname = os.path.join(self._tmppath, "test.test")
+        _fname = "test.test"
         EXP_IO_META.export_to_file(_fname)
         self.assertTrue(TestIo.exported)
         self.assertEqual(TestIo.export_filename, _fname)
 
-    def test_import_from_file(self):
-        _fname = os.path.join(self._tmppath, "test.test")
+    def test_import_from_file__generic(self):
+        _fname = "test.test"
         EXP_IO_META.import_from_file(_fname)
         self.assertTrue(TestIo.imported)
         self.assertEqual(TestIo.import_filename, _fname)
+        self.assertEqual(TestIo.diffraction_exp, EXP)
+
+    def test_import_from_file__given_Exp(self):
+        _exp = DiffractionExperiment()
+        _fname = "test.test"
+        EXP_IO_META.import_from_file(_fname, diffraction_exp=_exp)
+        self.assertTrue(TestIo.imported)
+        self.assertEqual(TestIo.import_filename, _fname)
+        self.assertEqual(TestIo.diffraction_exp, _exp)
 
 
 if __name__ == "__main__":